+++ test/test_application.py
@@ -1497,7 +1497,6 @@ async def create_visualisation_echo(name):
         template = VisualisationTemplate.objects.create(
             name="{name}",
             host_exact="{name}",
-            host_pattern="{name}",
             nice_name="Test {name}",
             spawner="PROCESS",
             spawner_options='{{"CMD":["python3", "/test/echo_server.py"]}}',
@@ -1528,7 +1527,6 @@ async def create_visualisation_dataset(name):
         template = VisualisationTemplate.objects.create(
             name="{name}",
             host_exact="{name}",
-            host_pattern="{name}",
             nice_name="Test {name}",
             spawner="PROCESS",
             spawner_options='{{"CMD":["python3", "/test/dataset_server.py"]}}',